// src/app/models/emprestimo.model.ts
export interface Emprestimo {
  id_emprestimo?: any | null;
  numero_conta?: any | null;
  nome_completo?: any | null;
  genero?: any | null;
  estado_civil?: any | null;
  dependentes?: any | null;
  educacao?: any | null;
  emprego?: any | null;
  propriedade?: any | null;
  credito?: any | null;
  renda_mensal_requerente?: any | null;
  renda_mensal_co_requerente?: any | null;
  valor_emprestimo?: any | null;
  duracao_emprestimo?: any | null;
  tipo_credito?: any | null;
  email?: any | null;
  resultado_predicao?: any | null;
}